slider bug on trending brands fixed
