+++ src/components/recommend/trending/TrendingFacetValues.jsx
@@ -10,6 +10,7 @@ import '@algolia/ui-components-horizontal-slider-theme';
 
 import { trendingConfig } from '@/config/trendingConfig';
 import CustomSkeleton from '@/components/skeletons/CustomSkeleton';
+import TrendingFacetsItem from '@/components/recommend/trending/TrendingFacetsItem';
 
 function WrappedTrendingFacetValues(props) {
   const { items, refine } = useRefinementList(props);
@@ -33,48 +34,6 @@ function WrappedTrendingFacetValues(props) {
     setRecommendationsLoaded(recommendations.length > 0);
   }, [recommendations]);
 
-  const TrendingFacetsItem = ({
-    trendingFacetValue
-  }) => {
-    // trendingFacet prop comes from Recommend, it is not a refinementList item, but we need a refinementList item to do things like refine.
-    // We look up the refinementList item which matches the current Recommend item (they are both facet values) and switch item.
-    // Item is now the refinementList item, so we can access all of correct functionality like isRefined etc.
-    const [isBusy, setBusy] = useState(true);
-    const [mergedItem, setMergedItem] = useState(null);
-
-    useEffect(() => {
-      if (items.length > 0) {
-        let newItems = items.filter(
-          (facet) => facet.label === trendingFacetValue.facetValue
-        );
-        setMergedItem(newItems[0]);
-        setBusy(false);
-      }
-    }, [items]);
-
-    return (
-      <>
-        {mergedItem && !isBusy && (
-          <button
-            className={`filters-container__content__list__button-filter ${mergedItem.isRefined ? 'refined-filter' : ''
-              }`}
-            type="button"
-            href="#"
-            onClick={(event) => {
-              event.preventDefault();
-              refine(mergedItem.value);
-            }}
-          >
-            <p>{mergedItem.label}</p>
-            <span className="filters-container__content__list__refinement-count">
-              {mergedItem.count}
-            </span>
-          </button>
-        )}
-      </>
-    );
-  };
-
   return (
     <div className="trending-facet-container">
       {recommendations.length > 0 && (
@@ -91,6 +50,7 @@ function WrappedTrendingFacetValues(props) {
                 <TrendingFacetsItem
                   trendingFacetValue={trendingFacetValue}
                   key={`${i}${trendingFacetValue}`}
+                  {...{items, refine}}
                 />
               ) : (
                 <div key={i + 'facetItem'}>
