+++ src/scss/index.scss
@@ -1,31 +1,5 @@
 @use './partials' as *;
 
-// Import Banners for the Home
-// @use '../components/banners/SCSS/banner.scss';
-// @use '../components/banners/SCSS/homeBanner2.scss';
-// @use '../components/banners/SCSS/homeBanner3.scss';
-// @use './banners/homeBanner.scss';
-// @use './banners/homeBanner2.scss';
-// @use './banners/homeBanner3.scss';
-
-// Import components
-// @use '../pages/productDetailsPage/SCSS/productDetails.scss';
-// @use '../components/appliedRules/SCSS/appliedRules.scss';
-// @use '../components/federatedSearch/SCSS/federatedSearch.scss';
-// @use '../components/footer/SCSS/footer.scss';
-// @use '../components/header/SCSS/header.scss';
-// @use '../components/demoGuide/SCSS/demoGuide.scss';
-// @use '../components/hits/SCSS/hits.scss';
-// @use '../components/hits/SCSS/skeleton.scss';
-// @use '../components/carousels/SCSS/carousels.scss';
-// @use '../components/loader/SCSS/loader.scss';
-// @use '../components/voicesearch/SCSS/voicesearch.scss';
-// @use '../components/searchbox/SCSS/searchbox.scss';
-// @use '../components/recommend/SCSS/recommend.scss';
-// @use '../components/facets/SCSS/facets.scss';
-// @use '../components/facets/SCSS/currentRefinement.scss';
-// @use '../pages/searchResultPage/SCSS/searchResultsPage.scss';
-
